venus_hfi: Use request_firmware_direct()
When firmware image for VIDC is absent in the filesystem the
kernel boot stuck for 60 seconds (being the default timeout).
Use the request_firmware_direct() is to take care of such aren't
always present firmware loading and hence replace the call in
the driver with this helper.

VIDC core init fault during kernel boot is acceptable because
it re-inited during each session open.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>